feat: 성공에 따른 UI 상태 처리 추가
